requirements.txt
python-dotenv==1.0.1
streamlit==1.43.2
openai==1.66.3